refactor(controllers): split KS + HR controllers into per-concern files
Pure cosmetic cleanup surfaced by the top-down architectural review's
god-object pass. Both controllers had ~500/400 LOC monolithic
controller.go files mixing four concerns:

  - lifecycle: Controller struct, Configure, Start, listener wiring
  - reconcile: the main reconcile body + helpers
  - dispatch: kind classifiers (shouldDispatchAsObject / isFluxSourceKind)
  - fingerprint: SHA256-of-spec dedup helper

The existing test files already split along these concerns
(dispatch_test.go, fingerprint_test.go, substitute_test.go for KS;
reconcile_test.go for HR). The implementation was the lone monolith.

This commit mirrors the test layout in the implementation files:

  pkg/controllers/kustomization/
    controller.go    (lifecycle + reconcile + collectDeps + resolveSourceRoot)
    fingerprint.go   (kustomizationFingerprint)
    substitute.go    (substituteDoc)
    dispatch.go      (emitRenderedChildren + helpers + kind classifiers)

  pkg/controllers/helmrelease/
    controller.go    (lifecycle + reconcile)
    fingerprint.go   (helmReleaseFingerprint + payload)
    dispatch.go      (isFluxSourceKind)

Net: −253 lines from monoliths, +254 across new files. controller.go
drops from 521→342 LOC (KS) and 380→315 LOC (HR), focused on the
reconcile narrative.

No behavior change. Identical imports re-distributed; package boundary
unchanged; all 5 test repos pass with identical numbers.

Verified:
- go test -count=1 ./...      — full suite green
- golangci-lint run ./...     — 0 issues
- tholinka/home-ops           — 118/0
- JJGadgets/Biohazard         — 199/0
- m00nwtchr/homelab-cluster   — 287/2 (pre-existing 404)
- buroa/k8s-gitops            — 73/0
- Zariel/home-ops             — 93/0
