+++ pkg/controllers/kustomization/controller.go
@@ -7,16 +7,10 @@ package kustomization
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"log/slog"
 	"path/filepath"
-	"strings"
 
-	kustomizev1 "github.com/fluxcd/kustomize-controller/api/v1"
-	yaml "go.yaml.in/yaml/v4"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/home-operations/flate/pkg/change"
@@ -249,185 +243,6 @@ func (c *Controller) reconcile(ctx context.Context, ks *manifest.Kustomization) 
 	})
 	return nil
 }
-
-// kustomizationFingerprint produces a stable hash of the post-Prepare
-// inputs that determine kustomize.RenderFlux's output for ks. The
-// resolved sourceRoot is included so a sibling KS that points at a
-// different source artifact path doesn't collide. Labels and
-// annotations are excluded on purpose: kustomize-controller-emitted
-// children carry stamped ownership labels that don't affect the
-// rendered manifests, and re-rendering on that delta is pure waste.
-// Returns "" when json.Marshal fails — empty fingerprints never
-// match, so the dedup short-circuit degrades safely into re-render.
-func kustomizationFingerprint(ks *manifest.Kustomization, sourceRoot string) string {
-	payload := struct {
-		Path                string
-		SourceRoot          string
-		Contents            map[string]any
-		PostBuildSubstitute map[string]any
-		Spec                kustomizev1.KustomizationSpec
-	}{
-		Path:                ks.Path,
-		SourceRoot:          sourceRoot,
-		Contents:            ks.Contents,
-		PostBuildSubstitute: ks.PostBuildSubstitute,
-		Spec:                ks.KustomizationSpec,
-	}
-	raw, err := json.Marshal(payload)
-	if err != nil {
-		return ""
-	}
-	sum := sha256.Sum256(raw)
-	return hex.EncodeToString(sum[:])
-}
-
-// emitRenderedChildren parses the rendered docs and lands them in the
-// store using a two-pass emission strategy:
-//
-//   - Pass 1 — non-leaf "data" kinds (ConfigMap, Secret, sources, etc.)
-//     go into the store first. Sources go through AddObject because
-//     they have their own status to track; ConfigMap/Secret have no
-//     controller so AddObject's event dispatch is a no-op for them.
-//     Either way they're in the store before pass 2 fires.
-//
-//   - Pass 2 — leaf reconcilables (Kustomization, HelmRelease). Their
-//     substituteFrom / chartRef lookups now see the data from pass 1.
-//
-// Without the two passes, AddObject for a reconcilable kind fires its
-// controller on a separate goroutine immediately, racing the parent's
-// "data first" emission. Within each pass the controller renders docs
-// in kustomize's emission order; passes themselves are ordered so the
-// data backing a reconcile always arrives first.
-//
-// Parse errors on inline resources are debug-logged and skipped — they
-// may be raw Kubernetes manifests flate doesn't track. SOPS-encrypted
-// secrets are debug-noted; ParseSecret wipes their values to the
-// PLACEHOLDER token the same way --wipe-secrets does for cleartext.
-func (c *Controller) emitRenderedChildren(id manifest.NamedResource, docs []map[string]any) {
-	type parsed struct {
-		obj          manifest.BaseManifest
-		reconcilable bool
-	}
-	opts := manifest.ParseDocOptions{WipeSecrets: c.WipeSecrets}
-	objs := make([]parsed, 0, len(docs))
-	for _, doc := range docs {
-		if manifest.IsEncryptedSecret(doc) {
-			name, ns := manifest.DocMetadata(doc)
-			slog.Debug("kustomization: SOPS-encrypted resource wiped to placeholder",
-				"id", id.String(), "ref", manifest.DocKind(doc)+" "+ns+"/"+name)
-		}
-		obj, err := manifest.ParseDoc(doc, opts)
-		if err != nil {
-			slog.Debug("kustomization: skipped doc", "id", id.String(), "err", err)
-			continue
-		}
-		objs = append(objs, parsed{obj: obj, reconcilable: shouldDispatchAsObject(obj)})
-	}
-	// Pass 1 — data first.
-	for _, p := range objs {
-		if p.reconcilable && isLeafReconcilable(p.obj) {
-			continue
-		}
-		if p.reconcilable {
-			c.keepEmitted(p.obj.Named())
-			c.Store.AddObject(p.obj)
-			c.markRendered(id, p.obj.Named())
-		} else {
-			c.Store.AddRendered(p.obj)
-		}
-	}
-	// Pass 2 — leaf reconcilables.
-	for _, p := range objs {
-		if p.reconcilable && isLeafReconcilable(p.obj) {
-			c.keepEmitted(p.obj.Named())
-			c.Store.AddObject(p.obj)
-			c.markRendered(id, p.obj.Named())
-		}
-	}
-}
-
-// keepEmitted extends the change filter's keep set so render-emitted
-// children pass the changed-only-mode PreGate check. Without this,
-// kustomize component+replacement patterns (parent KS emitting a
-// per-app KS via ConfigMap-driven replacements) produce silent gaps
-// in `flate diff`: the leaf KS isn't keep-set'd at filter-build time
-// because it didn't exist on disk, never reconciles, and its render
-// output never reaches the diff comparison. Issue #204.
-//
-// Called BEFORE Store.AddObject so the listener that fires
-// synchronously during AddObject sees the extended keep set when it
-// invokes PreGate.
-func (c *Controller) keepEmitted(id manifest.NamedResource) {
-	if f := c.Filter(); f != nil {
-		f.Add(id)
-	}
-}
-
-// substituteDoc marshals a single manifest doc, runs envsubst over it,
-// and unmarshals the result back. Per-doc substitution (rather than
-// substitute-the-whole-blob) lets us honor Flux's
-// "kustomize.toolkit.fluxcd.io/substitute: disabled" opt-out, which is
-// scoped to individual resources. The marshal/unmarshal round-trip is
-// load-bearing — it preserves Flux's YAML type-coercion semantics where
-// `replicas: ${REPLICAS}` (plain scalar) round-trips through envsubst
-// as int rather than string. Cheap pre-check on the decoded tree skips
-// the round-trip for the (common) case of docs with no `${` anywhere.
-func substituteDoc(doc map[string]any, vars map[string]string) (map[string]any, error) {
-	if !manifest.AnyStringLeaf(doc, func(s string) bool { return strings.Contains(s, "${") }) {
-		return doc, nil
-	}
-	raw, err := yaml.Marshal(doc)
-	if err != nil {
-		return nil, fmt.Errorf("substitute: marshal doc: %w", err)
-	}
-	out, err := kustomize.Substitute(raw, vars)
-	if err != nil {
-		return nil, err
-	}
-	var next map[string]any
-	if err := yaml.Unmarshal(out, &next); err != nil {
-		return nil, fmt.Errorf("substitute: unmarshal doc: %w", err)
-	}
-	return next, nil
-}
-
-// shouldDispatchAsObject reports whether a render-emitted Flux
-// resource needs to fire EventObjectAdded so its own controller picks
-// it up. The pattern is: parent Kustomization renders → emits a
-// child Flux resource (e.g. another Kustomization with parent patches
-// applied, a HelmRelease, an OCIRepository fanned out from a kustomize
-// component) → that child's controller must reconcile the patched
-// version, not the statically-loaded one.
-func shouldDispatchAsObject(obj manifest.BaseManifest) bool {
-	switch obj.(type) {
-	case *manifest.Kustomization,
-		*manifest.HelmRelease,
-		*manifest.HelmRepository,
-		*manifest.OCIRepository,
-		*manifest.GitRepository,
-		*manifest.Bucket,
-		*manifest.HelmChartSource,
-		*manifest.ExternalArtifact,
-		*manifest.ConfigMap,
-		*manifest.Secret:
-		return true
-	}
-	return false
-}
-
-// isLeafReconcilable reports whether an emitted object should be held
-// for pass 2. Kustomization + HelmRelease have controllers that fire
-// substituteFrom / chartRef lookups against the store the instant
-// their AddObject event arrives; emitting them after all "data" kinds
-// guarantees those lookups succeed.
-func isLeafReconcilable(obj manifest.BaseManifest) bool {
-	switch obj.(type) {
-	case *manifest.Kustomization, *manifest.HelmRelease:
-		return true
-	}
-	return false
-}
-
 // collectDeps assembles the dependency refs whose readiness must
 // precede this Kustomization: explicit dependsOn entries (carrying any
 // CEL ReadyExpr), the source ref, the implicit structural parent (the
